requirements.txt
accelerate==0.30.1
cv_bridge==1.13.0.post0
diffusers==0.32.2
external==0.0.1
fire==0.7.1
huggingface_hub==0.24.5
liger_kernel==0.6.1
matplotlib==3.9.4
numpy==1.26.4
opencv_python==4.11.0.86
Pillow==10.4.0
Requests==2.32.5
safetensors==0.6.2
timm==1.0.22
tqdm==4.67.1
transformers==4.43.3
einops==0.8.0

deepspeed==0.18.8
datasets==4.0.0
omegaconf==2.3.0